Move redirect on error to root layout
